+++ R/GenomicDataStream.R
@@ -13,7 +13,19 @@ NULL
 #' Read genomic data files (VCF, BCF, BGEN, h5ad) into R/Rcpp in chunks for analysis with Armadillo or Eigen libraries
 #'
 #' @export
-setClass("GenomicDataStream", slots = list(initialized = "logical", ptr = "externalptr", file = "character", field = "character", region = "character", samples = "character", chunkSize = "integer", missingToMean = "logical", featuresRead = "integer", streamType = "character", nsamples = "integer"))
+setClass("GenomicDataStream", slots = list(
+  initialized = "logical", 
+  ptr = "externalptr", 
+  file = "character", 
+  field = "character", 
+  region = "character", 
+  samples = "character", 
+  chunkSize = "integer", 
+  missingToMean = "logical", 
+  featuresRead = "integer", 
+  streamType = "character", 
+  nsamples = "integer", 
+  minVariance = "numeric"))
 
 
 #' @export
@@ -36,9 +48,13 @@ as.list.GenomicDataStream <- function(x, ...) {
 #' @param field field of VCF/BCF to read
 #' @param region target in the format \code{chr2:1-12345}. Multiple regions can be separated by one of \code{",\n\t"}, for example \code{"chr2:1-12345, chr3:1000-8000"}. Setting region to \code{""} includes all variants
 #' @param samples string of comma separated sample IDs to extract: \code{"ID1,ID2,ID3"}.  \code{"-"} indicates all samples
+#' @param MAF generalized minor allele frequency cutoff keeps variants with variance > \eqn{2(1-f)f}.
+#' @param minVariance features with variance \code{>= minVariance} are retained.  Defaults to \eqn{2(1-f)f}. If \code{NaN}, no filtering is applied.
 #' @param chunkSize	number of variants to return per chunk
 #' @param missingToMean	if true, set missing values to the mean dosage value. if false, set to \code{NaN}
 #' @param initialize default \code{FALSE}.  If \code{TRUE}, file info is read from path, otherwise store path until \code{GenomicDataStream} is initialized later
+#'
+#' @details Consider minor allele frequency (MAF) \eqn{f} and  Hardy-Weinberg equilibrium, the allelic states have probability \eqn{f^2, 2f(1-f), (1-f)^2}. If the variant has mean \eqn{\mu} and variance \eqn{\sigma^2}, MAF can be estimated from the mean as \eqn{min(\mu/2, 1 - \mu/2)} or from the variance as \eqn{p = 1+sqrt(1-2\sigma^2))/2} and MAF = \eqn{min(p, 1-p)}.  In addition the sample variance of the variant is \eqn{2(1-f)f}.  Therefore, setting a MAF cutoff corresponds to a variance cutoff that can also apply to multi-allelic variants.
 #'
 #' @return object of class \code{GenomicDataStream}
 #'
@@ -64,15 +80,28 @@ as.list.GenomicDataStream <- function(x, ...) {
 #' #
 #' @importFrom methods new is
 #' @export
-GenomicDataStream <- function(file, field = "", region = "", samples = "-", chunkSize = 1000, missingToMean = FALSE, initialize = FALSE) {
+GenomicDataStream <- function(file, field = "", region = "", samples = "-", MAF = 0, minVariance = 2*(1-MAF)*MAF, chunkSize = 10000, missingToMean = TRUE, initialize = FALSE){
+
   chunkSize <- as.integer(chunkSize)
   samples <- paste(samples, collapse = ",")
 
   file = path.expand(file)
 
+  # check that file exists
+  if( ! file.exists(file) ){
+    stop("File does not exist")
+  }
+
+  if( MAF < 0 ){
+    stop("MAF must be >= 0")
+  }
+  if( !is.nan(minVariance) & minVariance < 0 ){
+    stop("minVariance must be >= 0")
+  }
+
   if (initialize) {
     # Create GenomicDataStream and return external pointer
-    ptr <- create_xptr(file, field, region, samples, chunkSize, missingToMean)
+    ptr <- create_xptr(file, field, region, samples, minVariance, chunkSize, missingToMean)
 
     # get additional information about data
     info <- getInfo(ptr)
@@ -85,6 +114,7 @@ GenomicDataStream <- function(file, field = "", region = "", samples = "-", chun
       field = field,
       region = region,
       samples = samples,
+      minVariance = minVariance,
       chunkSize = chunkSize,
       missingToMean = missingToMean,
       streamType = info$streamType,
@@ -98,6 +128,7 @@ GenomicDataStream <- function(file, field = "", region = "", samples = "-", chun
       field = field,
       region = region,
       samples = samples,
+      minVariance = minVariance,
       chunkSize = chunkSize,
       missingToMean = missingToMean
     )
@@ -166,15 +197,106 @@ initializeStream <- function(x) {
     field = x@field,
     region = x@region,
     samples = x@samples,
+    minVariance = x@minVariance,
     chunkSize = x@chunkSize,
     missingToMean = x@missingToMean,
     initialize = TRUE
   )
 }
 
-#' Detected if end of stream is reaached
+
+#' Set regions of GenomicDataStream
 #'
-#' Detected if end of stream is reaached
+#' Set regions of GenomicDataStream
+#'
+#' @param x \code{GenomicDataStream}
+#' @param region target in the format \code{chr2:1-12345}. Multiple regions can be separated by one of \code{",\n\t"}, for example \code{"chr2:1-12345, chr3:1000-8000"}. Setting region to \code{""} includes all variants
+#'
+#' @return \code{GenomicDataStream} with region set
+#'
+#' @description If \code{GenomicDataStream} is already initialized, set new query in C++ backend. Otherwise substitute \code{region} values
+#'
+#' @examples
+#' file <- system.file("extdata", "test.vcf.gz", package = "GenomicDataStream")
+#'
+#' obj <- GenomicDataStream(file, "DS", chunkSize = 5)
+#'
+#' # by default, GenomicDataStream is not initialized
+#' setRegion(obj, "1:10000-12000")
+#
+#' @export
+setRegion <- function(x, region) {
+  
+  if ( isInitialized(x) ) {
+    ptr <- setRegions_rcpp(x@ptr, region)
+
+    # get additional information about data
+    info <- getInfo(ptr)
+
+    obj <- new("GenomicDataStream",
+        initialized = TRUE,
+        ptr = ptr,
+        file = x@file,
+        field = x@field,
+        region = region,
+        samples = x@samples,
+        minVariance = x@minVariance,
+        chunkSize = x@chunkSize,
+        missingToMean = x@missingToMean,
+        streamType = info$streamType,
+        nsamples = info$nsamples)
+  }else{
+    obj <- new("GenomicDataStream",
+      initialized = FALSE,
+      file = x@file,
+      field = x@field,
+      region = region,
+      samples = x@samples,
+      minVariance = x@minVariance,
+      chunkSize = x@chunkSize,
+      missingToMean = x@missingToMean)
+  }
+
+  obj 
+}
+
+
+
+#' Get sample names 
+#'
+#' Get sample names in order that the genotypes are extracted
+#'
+#' @param x \code{GenomicDataStream}
+#'
+#' @return array of string names
+#'
+#' @description BGEN uses sample order from the query, but VCF/BCF/PGEN uses order in file
+#'
+#' @examples
+#' file <- system.file("extdata", "test.vcf.gz", package = "GenomicDataStream")
+#'
+#' obj <- GenomicDataStream(file, "DS", init=TRUE)
+#'
+#' getSampleNames(obj)
+#
+#' @export
+getSampleNames <- function(x) {
+  
+  if ( ! isInitialized(x) ) {
+    stop("Must be initialized first")
+  }
+
+  return( getSampleNames_rcpp( x@ptr ) )
+}
+
+
+
+
+
+
+#' Detected if end of stream is reached
+#'
+#' Detected if end of stream is reached
 #'
 #' @param x \code{GenomicDataStream}
 #'
@@ -298,6 +420,7 @@ setMethod(
       cat("  field:        ", object@field, "\n")
       cat("  region:       ", object@region, "\n")
       cat("  samples:      ", object@nsamples, "\n")
+      cat("  minVar cutoff:", object@minVariance, "\n")
       cat("  missingToMean:", object@missingToMean, "\n")
       cat("  chunkSize:    ", object@chunkSize, "\n")
       cat("  features read:", featuresRead(object), "\n")