+++ docs/doxygen/html/search/all_5.js
@@ -9,5 +9,6 @@ var searchData=
   ['datatable_2eh_6',['DataTable.h',['../_data_table_8h.html',1,'']]],
   ['delayedstream_7',['DelayedStream',['../classgds_1_1_delayed_stream.html',1,'gds::DelayedStream'],['../classgds_1_1_delayed_stream.html#a8c2cbc77c84dbf62c4dd7e78dc6a0034',1,'gds::DelayedStream::DelayedStream()']]],
   ['delayedstream_2eh_8',['DelayedStream.h',['../_delayed_stream_8h.html',1,'']]],
-  ['dependencies_9',['Dependencies',['../index.html#autotoc_md2',1,'']]]
+  ['dependencies_9',['Dependencies',['../index.html#autotoc_md2',1,'']]],
+  ['dst_10',['dst',['../struct_z_s_t_d__out_buffer__s.html#ae17b7e44abe8ffd1f72f448df86f2a23',1,'ZSTD_outBuffer_s']]]
 ];