+++ docs/doxygen/html/search/all_e.js
@@ -3,5 +3,6 @@ var searchData=
   ['r_0',['A scalable interface between data and analysis underneath R',['../index.html',1,'']]],
   ['r_5finit_5fgenomicdatastream_1',['R_init_GenomicDataStream',['../_rcpp_exports_8cpp.html#a28813cc54257bc2fa1103e5290158da5',1,'RcppExports.cpp']]],
   ['rcppexports_2ecpp_2',['RcppExports.cpp',['../_rcpp_exports_8cpp.html',1,'']]],
-  ['regions_3',['regions',['../structgds_1_1_param.html#a15532b76e72f2f03fd4336e99585eb4b',1,'gds::Param']]]
+  ['regions_3',['regions',['../structgds_1_1_param.html#a15532b76e72f2f03fd4336e99585eb4b',1,'gds::Param']]],
+  ['retainvariants_4',['retainVariants',['../classgds_1_1_variant_info.html#a95aa9d6aa717eed7ab7f921ff93311d5',1,'gds::VariantInfo']]]
 ];