+++ docs/doxygen/html/search/defines_2.js
@@ -1,0 +1,12 @@
+var searchData=
+[
+  ['zstd_5fexpand_5fand_5fquote_0',['ZSTD_EXPAND_AND_QUOTE',['../zstd_8h.html#a5f28056060ce509b3c9c7c6551a6ac28',1,'zstd.h']]],
+  ['zstd_5flib_5fversion_1',['ZSTD_LIB_VERSION',['../zstd_8h.html#a48024c26f41157f9f013bda534ad41e6',1,'zstd.h']]],
+  ['zstd_5fquote_2',['ZSTD_QUOTE',['../zstd_8h.html#adbbadef2056cbbc1b4d24523c2c6cb5b',1,'zstd.h']]],
+  ['zstd_5fversion_5fmajor_3',['ZSTD_VERSION_MAJOR',['../zstd_8h.html#aa4a867002655c8077460a47e5960e4bc',1,'zstd.h']]],
+  ['zstd_5fversion_5fminor_4',['ZSTD_VERSION_MINOR',['../zstd_8h.html#ad96b66eabc8a315ce4c2b6c34b29746f',1,'zstd.h']]],
+  ['zstd_5fversion_5fnumber_5',['ZSTD_VERSION_NUMBER',['../zstd_8h.html#a50ac5656bb6539eae98593afd3493132',1,'zstd.h']]],
+  ['zstd_5fversion_5frelease_6',['ZSTD_VERSION_RELEASE',['../zstd_8h.html#a82ca36516a6cfafe730e00f6e5cc3f23',1,'zstd.h']]],
+  ['zstd_5fversion_5fstring_7',['ZSTD_VERSION_STRING',['../zstd_8h.html#a930937e5d86408c3c866db9af7edd8b8',1,'zstd.h']]],
+  ['zstdlib_5fapi_8',['ZSTDLIB_API',['../zstd_8h.html#a9dbac0fbe1cd466cb2b82c5401e2ec49',1,'zstd.h']]]
+];