+++ docs/doxygen/html/search/functions_b.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['r_5finit_5fgenomicdatastream_0',['R_init_GenomicDataStream',['../_rcpp_exports_8cpp.html#a28813cc54257bc2fa1103e5290158da5',1,'RcppExports.cpp']]]
+  ['r_5finit_5fgenomicdatastream_0',['R_init_GenomicDataStream',['../_rcpp_exports_8cpp.html#a28813cc54257bc2fa1103e5290158da5',1,'RcppExports.cpp']]],
+  ['retainvariants_1',['retainVariants',['../classgds_1_1_variant_info.html#a95aa9d6aa717eed7ab7f921ff93311d5',1,'gds::VariantInfo']]]
 ];