+++ docs/doxygen/html/search/functions_c.js
@@ -1,8 +1,10 @@
 var searchData=
 [
   ['setfield_0',['setField',['../structgds_1_1_param.html#a0bedb8b7bc80339cdedb98066dbd157d',1,'gds::Param']]],
-  ['setrownames_1',['setRowNames',['../classgds_1_1_matrix_info.html#ae9531c5a2a3601465e47a504aa980b3e',1,'gds::MatrixInfo']]],
-  ['setsamplesfile_2',['setSamplesFile',['../structgds_1_1_param.html#a8c752be96dffc51b142914a8b2fe2f33',1,'gds::Param']]],
-  ['size_3',['size',['../classgds_1_1_data_info.html#af8d39001b9a18c530c6d60f5caf32789',1,'gds::DataInfo::size()'],['../classgds_1_1_genomic_ranges.html#a7d1b020b0a925df57b1d9b56ddbae8e3',1,'gds::GenomicRanges::size()']]],
-  ['standardize_5ftest_4',['standardize_test',['../export_8cpp.html#a7948cb70a34ad29b874afa09a0fd2558',1,'standardize_test(arma::mat &amp;X, const bool &amp;center=true, const bool &amp;scale=true):&#160;export.cpp'],['../_rcpp_exports_8cpp.html#a62357b90027c8c7431e79c4955b63428',1,'standardize_test(arma::mat &amp;X, const bool &amp;center, const bool &amp;scale):&#160;export.cpp']]]
+  ['setregions_1',['setRegions',['../classgds_1_1bgenstream.html#a951a5ee78e734d54a0b3ed63e60cd808',1,'gds::bgenstream::setRegions()'],['../classgds_1_1_delayed_stream.html#a323477c8f2ecfa8857477cfc1ca4afe3',1,'gds::DelayedStream::setRegions()'],['../classgds_1_1_genomic_data_stream.html#a5abd183e6738fa4ccc6609c0b7f63cc8',1,'gds::GenomicDataStream::setRegions()'],['../classgds_1_1pgenstream.html#ab43788ca0954f7b617aed631748f3179',1,'gds::pgenstream::setRegions()'],['../classgds_1_1vcfstream.html#ae1534b46ee3f3ea31c547b9c2ed554a1',1,'gds::vcfstream::setRegions()']]],
+  ['setregions_5frcpp_2',['setRegions_rcpp',['../export_8cpp.html#a2dd3e4034f7abff758612d5c610d7930',1,'setRegions_rcpp(SEXP x, const string &amp;regionString):&#160;export.cpp'],['../_rcpp_exports_8cpp.html#a2dd3e4034f7abff758612d5c610d7930',1,'setRegions_rcpp(SEXP x, const string &amp;regionString):&#160;export.cpp']]],
+  ['setrownames_3',['setRowNames',['../classgds_1_1_matrix_info.html#ae9531c5a2a3601465e47a504aa980b3e',1,'gds::MatrixInfo']]],
+  ['setsamplesfile_4',['setSamplesFile',['../structgds_1_1_param.html#a8c752be96dffc51b142914a8b2fe2f33',1,'gds::Param']]],
+  ['size_5',['size',['../classgds_1_1_data_info.html#af8d39001b9a18c530c6d60f5caf32789',1,'gds::DataInfo::size()'],['../classgds_1_1_genomic_ranges.html#a7d1b020b0a925df57b1d9b56ddbae8e3',1,'gds::GenomicRanges::size()']]],
+  ['standardize_5ftest_6',['standardize_test',['../export_8cpp.html#a7948cb70a34ad29b874afa09a0fd2558',1,'standardize_test(arma::mat &amp;X, const bool &amp;center=true, const bool &amp;scale=true):&#160;export.cpp'],['../_rcpp_exports_8cpp.html#a62357b90027c8c7431e79c4955b63428',1,'standardize_test(arma::mat &amp;X, const bool &amp;center, const bool &amp;scale):&#160;export.cpp']]]
 ];