+++ docs/doxygen/html/search/variables_2.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['data_0',['data',['../classgds_1_1_data_chunk.html#ad98671d009e299f2a5e015677431d98b',1,'gds::DataChunk']]]
+  ['data_0',['data',['../classgds_1_1_data_chunk.html#ad98671d009e299f2a5e015677431d98b',1,'gds::DataChunk']]],
+  ['dst_1',['dst',['../struct_z_s_t_d__out_buffer__s.html#ae17b7e44abe8ffd1f72f448df86f2a23',1,'ZSTD_outBuffer_s']]]
 ];