+++ docs/doxygen/html/search/variables_3.js
@@ -3,6 +3,7 @@ var searchData=
   ['featuresread_0',['featuresRead',['../structgds_1_1_bound_data_stream.html#a07ecefe29d25a50bf1b0f980479d6f48',1,'gds::BoundDataStream']]],
   ['field_1',['field',['../structgds_1_1_param.html#aad1725c672bbad45c1ea52f6c9593da6',1,'gds::Param']]],
   ['file_2',['file',['../structgds_1_1_param.html#ada92631b1e8dfaec19d59cb4225540c5',1,'gds::Param']]],
-  ['filesamples_3',['fileSamples',['../structgds_1_1_param.html#a83c54c3a294a8eb8b22458f6f0780a82',1,'gds::Param']]],
-  ['filetype_4',['fileType',['../structgds_1_1_param.html#a8772547522722593acf7f4f96cb113ff',1,'gds::Param']]]
+  ['filenameidxglobal_3',['filenameIdxGlobal',['../classgds_1_1bgenstream.html#a98ed5d54dfde37dd1e573c750a25777a',1,'gds::bgenstream']]],
+  ['filesamples_4',['fileSamples',['../structgds_1_1_param.html#a83c54c3a294a8eb8b22458f6f0780a82',1,'gds::Param']]],
+  ['filetype_5',['fileType',['../structgds_1_1_param.html#a8772547522722593acf7f4f96cb113ff',1,'gds::Param']]]
 ];