+++ docs/doxygen/html/search/variables_5.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['missingtomean_0',['missingToMean',['../structgds_1_1_param.html#af5eae8c29ee634c69c59e5f76d8df801',1,'gds::Param']]]
+  ['minvariance_0',['minVariance',['../structgds_1_1_param.html#a90139c51a3a8a7f14e32d7404efa44dd',1,'gds::Param']]],
+  ['missingtomean_1',['missingToMean',['../structgds_1_1_param.html#af5eae8c29ee634c69c59e5f76d8df801',1,'gds::Param']]]
 ];