+++ docs/doxygen/html/search/variables_6.js
@@ -2,6 +2,7 @@ var searchData=
 [
   ['param_0',['param',['../classgds_1_1_genomic_data_stream.html#a707058d3851890ff8da01e5485de164a',1,'gds::GenomicDataStream']]],
   ['pos_1',['POS',['../classgds_1_1_variant_info.html#ab14ce9b79f6a06ea228606f90810f543',1,'gds::VariantInfo']]],
-  ['position_2',['position',['../structgds_1_1point.html#adc7cb6d1663aac15a58ab5720b36af59',1,'gds::point']]],
-  ['ptr_3',['ptr',['../structgds_1_1_bound_data_stream.html#abf1a91c6f22f65fbd200b618fc8180f9',1,'gds::BoundDataStream']]]
+  ['pos_2',['pos',['../struct_z_s_t_d__in_buffer__s.html#af25828e77def7d36cb648494649ecaa4',1,'ZSTD_inBuffer_s::pos'],['../struct_z_s_t_d__out_buffer__s.html#a51b01c063d6e338d333a865b88ea243c',1,'ZSTD_outBuffer_s::pos']]],
+  ['position_3',['position',['../structgds_1_1point.html#adc7cb6d1663aac15a58ab5720b36af59',1,'gds::point']]],
+  ['ptr_4',['ptr',['../structgds_1_1_bound_data_stream.html#abf1a91c6f22f65fbd200b618fc8180f9',1,'gds::BoundDataStream']]]
 ];