+++ docs/doxygen/html/search/variables_8.js
@@ -1,5 +1,7 @@
 var searchData=
 [
   ['samplenames_0',['sampleNames',['../classgds_1_1_variant_info.html#ad77368d79bf9efcaf87fd004c7a41efa',1,'gds::VariantInfo']]],
-  ['samples_1',['samples',['../structgds_1_1_param.html#a73e7ac933fc1c6305261e9179eac9732',1,'gds::Param']]]
+  ['samples_1',['samples',['../structgds_1_1_param.html#a73e7ac933fc1c6305261e9179eac9732',1,'gds::Param']]],
+  ['size_2',['size',['../struct_z_s_t_d__in_buffer__s.html#ad4673a0122dbe52fe6b0aa9824a35c29',1,'ZSTD_inBuffer_s::size'],['../struct_z_s_t_d__out_buffer__s.html#a34fb461ecc114951093b978c3aa6899e',1,'ZSTD_outBuffer_s::size']]],
+  ['src_3',['src',['../struct_z_s_t_d__in_buffer__s.html#a2a838d5f348ddfa781bb14aa5209e9eb',1,'ZSTD_inBuffer_s']]]
 ];