+++ inst/include/utils.h
@@ -1,6 +1,4 @@
 
-
-
 #include <armadillo>
 
 #include <vector>
@@ -9,10 +7,10 @@
 #include <unordered_set>
 #include <regex>
 
+#include <boost/algorithm/string.hpp>
 
 #ifndef UTILS_H_
 #define UTILS_H_
-
 
 namespace gds {
 
@@ -266,7 +264,45 @@ vector<T> cast_elements( const vector<string> &v ){
 }
 
 
+/** regionString is string of chr:start-end delim by "\t,\n"
+ remove spaces, then split based on delim
+ remove duplicate regions, but preserve order
+ Note: regionString is copy by value, since boost::erase_all overwrites
+ */
+static vector<string> splitRegionString( string regionString){
+
+    vector<string> regions;
+
+    // regionString is string of chr:start-end delim by "\t,\n"
+    // remove spaces, then split based on delim
+    boost::erase_all(regionString, " ");
+    boost::split(regions, regionString, boost::is_any_of("\t,\n"));
+
+    // remove duplicate regions, but preserve order
+    removeDuplicates( regions );
+
+    return regions;
 }
 
 
+/** Return subset of x indicated by entries in ind: x[idx] 
+*/
+template<typename T>
+static vector<T> subset_vector(const vector<T> &x, const vector<unsigned int> &idx){
+
+    // initialize x_subset to have size idx.size()
+    vector<T> x_subset;
+    x_subset.reserve(idx.size());
+
+    for(int i=0; i<idx.size(); i++){
+        if( idx[i] > x.size() ){
+            throw std::out_of_range("Index is out of bounds"); 
+        }
+        x_subset.push_back(x[idx[i]]);
+    }
+
+    return x_subset;
+}
+
+} // end namespace
 #endif